Add examples navigation (#10)
* Release v0.0.2+1 (#9)

* Implement tag selector component (#2)

* Feat/implement accordions single date picker and range date picker (#3)

* Implement accordion, single date picker, date range picker

* Manage responsive on date picker

* Make weekStartsOn fonctionnal on date Picker

* add maxSelectableRange on date range picker

* Add join selected days on date range picker

* Implement multiform step component (#4)

* Implement custom notification component (#6)

* Implement advanced pagination with example (#5)

* Implement advanced pagination with example

* Change files location

* Add load more button

* Resolve merge conflicts

* Resolve merge conflicts

* Update readme file

* Update MD files for deployment

* Update logo link

* Update logo link

* Update package version

* Resolde dart a anlysis issues

* ✨ feat: add HomePage with example navigation

Introduce a new HomePage that serves as a central hub for 
navigating various examples in the Flu Kit. The HomePage 
displays a list of examples, allowing users to tap and 
navigate to each specific example. This change enhances 
the user experience by providing a structured way to 
access different components of the library.

* ✨ feat: add AppBar to QueedNotificationExample widget

Adds an AppBar with a title to the QueedNotificationExample 
widget for improved navigation and context. This enhances 
the user interface by providing a clear title for the 
example screen.

---------

Co-authored-by: maeltoukap <74214399+maeltoukap@users.noreply.github.com>